CLRadeonExtender: Fixed gcc warning for 32-bit compilation.

+++ amdasm/GCNAsmHelpers.cpp
@@ -147,8 +147,8 @@ bool GCNAsmUtils::parseRegVarRange(Assembler& asmr, const char*& linePtr,
             }
             
             if (regField!=ASMFIELD_NONE)
-                section.addVarUsage({ asmr.currentOutPos, regField, uint16_t(rstart),
-                    uint16_t(rend), (flags & INSTROP_READ)!=0,
+                section.addVarUsage({ size_t(asmr.currentOutPos), regField,
+                    uint16_t(rstart), uint16_t(rend), (flags & INSTROP_READ)!=0,
                     (flags & INSTROP_WRITE)!=0, 0, regVar });
             return true;
         }
